pyproject.toml
[project]
name = "upload-service"
version = "0.1.0"
description = "A production-grade S3 upload service with parallel upload support"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "boto3>=1.26.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "pytest>=7.4.0",
    "moto>=4.2.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 